requirements.txt
tensorflow
pillow
flask
flask-cors
opencv-python
